backend/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Authentication & Security
PyJWT==2.10.1
cryptography==44.0.0
passlib==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.18

# AI/ML Stack (Python 3.13 compatible versions)
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.12
openai==1.58.1
pinecone-client==6.0.0
voyageai==0.3.4

# Document Processing
pypdf2==3.0.1
python-docx==1.1.2
pillow==11.1.0

# File Storage
boto3==1.35.89
python-dotenv==1.0.1

# HTTP Client
httpx==0.28.1
aiofiles==24.1.0

# Development
pytest==8.3.4
pytest-asyncio==0.25.0
